Add files via upload
bl4ckbo7 Password Generator is a native C++ application that is built up with functionalities that help the user to generate randomly passwords of their choice. This passwords MUST range from atleast 8 characters (which has to be the minimum password length), thus from there the user has the choice to generate different combinations of characters and numbers and symbols in his/her password.